pyproject.toml
[project]
name = "todoproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.17.1",
]
